Refactoring, fixes
