+++ Assets/Scripts/Learning/AbstractAcademy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60ba6fd8918e8bf43a1639785f88ea0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
