+++ Assets/Scripts/NeuroEvolution/Interfaces/IGeneticController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54d779d41bd6eed4095c2de318738df0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
