+++ Assets/Scripts/UI/Abstract_UI_Genetics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f98e1727492eb5246a8de2d770f0f90e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
