+++ Assets/Scripts/UI/UI_Genetics_NEAT.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 083eef6d8eb338f4a8d12ebd5d85b42d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
